refactor: remove not needed import, move to better folder structure
