+++ apps/dokploy/components/dashboard/swarm/servers/servers-overview.tsx
@@ -1,5 +1,5 @@
 import { api } from "@/utils/api";
-import { ServerOverviewCard } from "../server-card";
+import { ServerOverviewCard } from "./server-card";
 
 export default function ServersOverview() {
 	const { data: servers, isLoading } = api.server.all.useQuery();
